Add files via upload
A Vegetable Website is a simple web project showcasing different vegetables with descriptions, images, and benefits. It is built using HTML and CSS  for structure, styling, and interactivityand ,php for the database connection of login details. The site can include a homepage, an about section, and a contact form. After development, the website is uploaded to GitHub and hosted using GitHub Pages for public access.