Update README.md
A Vegetable Website with User Login is a dynamic web project that allows users to explore vegetable-related content while having a secure login and registration system using PHP and MySQL. The website includes a homepage, user authentication (login, registration, logout), and a dashboard for logged-in users. The backend is built with PHP, and user data is securely stored in a MySQL database with password encryption. The project is hosted on GitHub and can be run locally using XAMPP or MAMP.